gitriders/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sovereign-export"
version = "1.0.0"
description = "GitRiders - FlameLang Sovereignty Export System for AI chat conversations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "StrategicKhaos DAO LLC", email = "legion@strategickhaos.com"}
]
keywords = ["ai", "sovereignty", "export", "encryption", "privacy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "cryptography>=41.0.0",
    "pynacl>=1.5.0",
    "requests>=2.31.0",
    "requests-oauthlib>=1.3.1",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "argon2-cffi>=23.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
sovereign-export = "sovereign_export.cli:main"

[project.urls]
Homepage = "https://github.com/gitriders/sovereign-export"
Documentation = "https://github.com/gitriders/sovereign-export/docs"
Repository = "https://github.com/gitriders/sovereign-export"
Issues = "https://github.com/gitriders/sovereign-export/issues"

[tool.setuptools]
packages = ["sovereign_export", "sovereign_export.connectors"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --tb=short"
